refactor(checker): extract enum evaluation utilities (#14351)
