+++ crates/tsz-checker/src/types/utilities/mod.rs
@@ -8,6 +8,7 @@ pub(crate) mod contextual_parameters;
 pub(crate) mod core;
 pub(crate) mod cycle_guard;
 pub(crate) mod element_indexable;
+pub(crate) mod enum_eval;
 pub(crate) mod enum_utils;
 pub(crate) mod enum_utils_readonly;
 pub(crate) mod heritage_walk_state;
